api/requirements.txt
fastapi>=0.110.0
anthropic>=0.40.0
libsql-experimental>=0.0.50
pyjwt>=2.8.0
bcrypt>=4.1.0
python-multipart>=0.0.9
httpx>=0.27.0
fpdf2>=2.7.0
